refactor : divide freetalk apis #48
